Move tests from execution::context
